refactor: 날짜 포맷 관련 함수 util 함수로 분리
